Rename base editor module and add required methods to base class
