+++ iterm2_dwim/editors/base.py
@@ -1,0 +1,10 @@
+class BaseEditor(object):
+    """
+    Abstract base class for editors.
+    """
+
+    def __init__(self, executable):
+        self.executable = executable
+
+    def visit_file(self, path, text_after):
+        raise NotImplementedError()
