--- iterm2_dwim/editors/base_editor.py
@@ -1,3 +1,0 @@
-class BaseEditor(object):
-    def __init__(self, executable):
-        self.executable = executable
